Remove styledsystemcomponents. Split components into components folder and add default imports
