+++ webapp/src/components/Box/Box.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { spacing, palette, typography, sizing } from '@material-ui/system';
+import Box from '@material-ui/core/Box';
+
+const StyledBox = styled(Box)`
+  ${spacing}
+  ${palette}
+`;
+
+export default StyledBox;
