+++ webapp/src/components/Link/Link.js
@@ -1,0 +1,9 @@
+import { Link } from 'react-router-dom';
+import styled from 'styled-components';
+
+const LinkWrapper = styled(Link)`
+  text-decoration: none;
+  color: #000;
+`;
+
+export default LinkWrapper;
