+++ webapp/src/components/index.js
@@ -1,0 +1,27 @@
+//* Main import files for all shared components
+
+import StyledBox from './Box/Box';
+import Card from './Card/Card';
+import {
+	StyledButtonPrimary,
+	StyledButtonSecondary,
+	StyledButtonDelete,
+	ButtonWrapper,
+	ButtonArea,
+	StyledButtonInteraction
+} from './Button/Button';
+import LinkWrapper from './Link/Link';
+import StyledTooltip from './Tooltip/Tooltip';
+
+export {
+	StyledBox,
+	Card,
+	StyledButtonPrimary,
+	StyledButtonSecondary,
+	StyledButtonDelete,
+	ButtonWrapper,
+	ButtonArea,
+	StyledButtonInteraction,
+	LinkWrapper,
+	StyledTooltip
+};
