+++ webapp/src/shared/CardWrapper.js
@@ -1,15 +1,13 @@
-import Card from "@material-ui/core/Card";
-import CardActionArea from "@material-ui/core/CardActionArea";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import CardMedia from "@material-ui/core/CardMedia";
-import styled from "styled-components";
-import { spacing, palette, typography } from "@material-ui/system";
+import Card from '@material-ui/core/Card';
+import CardActionArea from '@material-ui/core/CardActionArea';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import styled from 'styled-components';
+import { spacing, palette, typography } from '@material-ui/system';
 
-const IWrapper = styled.i`
-  ${spacing};
-  ${palette};
-`;
+// TODO: Need some refactor afterward.
+
 const CardWrapper = styled(Card)`
   ${spacing};
   ${palette};
@@ -32,12 +30,12 @@ const CardWrapperFullWidth = styled(Card)`
   height: 100%;
 `;
 const DivWrapperHalfWidth = styled.div`
-  ${spacing};
-  ${palette};
-  max-width: 50%;
-  display: flex;
-  justify-content: center;
-  border-radius: 15px;
+	${spacing};
+	${palette};
+	max-width: 50%;
+	display: flex;
+	justify-content: center;
+	border-radius: 15px;
 `;
 const CardActionsWrapper = styled(CardActions)`
   ${spacing};
@@ -65,14 +63,13 @@ const CardMediaWrapper = styled(CardMedia)`
 `;
 
 export {
-  IWrapper,
-  CardWrapper,
-  CardActionsWrapper,
-  CardActionAreaWrapper,
-  CardContentWrapper,
-  CardMediaWrapper,
-  CardWrapperFullWidth,
-  CardWrapperFlexColumn,
-  CardWrapperFlex,
-  DivWrapperHalfWidth,
+	CardWrapper,
+	CardActionsWrapper,
+	CardActionAreaWrapper,
+	CardContentWrapper,
+	CardMediaWrapper,
+	CardWrapperFullWidth,
+	CardWrapperFlexColumn,
+	CardWrapperFlex,
+	DivWrapperHalfWidth
 };
